requirements.txt
Django==5.0.4
djangorestframework==3.15.1
mysqlclient==2.2.7
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.1.1
